Message Filter - Initial Commit
Main objective of creating this application is to filter incoming SMS from
various promoters. Most of these messages contain a Name (Text) as the sender
instead of a phone number, this cause most of the similar applications out
there to not be able to filter by the sender.

This commit contain the core function of the application, which is intercept
the incoming SMS via Android Broadcast and display a notification.Filtration
and stats will be added prior to the first release.