Database and stat recording
Adding the base schema, helper and the data layer classes for the main data
store of the application.

Broadcast receiver will be recording each messages sender and time stamp to the
messages table to be used in the stat reports later.

Blocked sender data storage is available but is not integrated to the filter
functionality yet.