Adding the actual "block" function
SMS broadcast receiver will lookup the blocked senders list and if the send is
in there, it will stop the broadcast from propagating.

Trick here is that the priority of this broadcast receiver's intent filter is
to the maximum (2^31 - 1). It is not determined how this will behave when a
application with same priority listens to the same intent.